refactor: move tests in dedicated file
