+++ tests/test_main.py
@@ -1,18 +1,9 @@
 import sys
-from collections import namedtuple
 from enum import Enum
-from typing import Any, Callable, ClassVar, Dict, List, Mapping, NamedTuple, Optional, Type, get_type_hints
+from typing import Any, Callable, ClassVar, Dict, List, Mapping, Optional, Type, get_type_hints
 from uuid import UUID, uuid4
 
 import pytest
-
-if sys.version_info < (3, 8):
-    try:
-        from typing import TypedDict
-    except ImportError:
-        TypedDict = None
-else:
-    from typing import TypedDict
 
 from pydantic import (
     BaseModel,
@@ -1434,90 +1425,3 @@ def test_base_config_type_hinting():
         a: int
 
     get_type_hints(M.__config__)
-
-
-def test_named_tuple():
-    Position = namedtuple('Pos', 'x y')
-
-    class Event(NamedTuple):
-        a: int
-        b: int
-        c: int
-        d: str
-
-    class Model(BaseModel):
-        pos: Position
-        events: List[Event]
-
-    model = Model(pos=('1', 2), events=[[b'1', '2', 3, 'qwe']])
-    assert isinstance(model.pos, Position)
-    assert isinstance(model.events[0], Event)
-    assert model.pos.x == '1'
-    assert model.pos == Position('1', 2)
-    assert model.events[0] == Event(1, 2, 3, 'qwe')
-    assert repr(model) == "Model(pos=Pos(x='1', y=2), events=[Event(a=1, b=2, c=3, d='qwe')])"
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(pos=('1', 2), events=[['qwe', '2', 3, 'qwe']])
-    assert exc_info.value.errors() == [
-        {
-            'loc': ('events', 0, 'a'),
-            'msg': 'value is not a valid integer',
-            'type': 'type_error.integer',
-        }
-    ]
-
-
-@pytest.mark.skipif(not TypedDict, reason='typing_extensions not installed')
-def test_typed_dict():
-    class TD(TypedDict):
-        a: int
-        b: int
-        c: int
-        d: str
-
-    class Model(BaseModel):
-        td: TD
-
-    m = Model(td={'a': '3', 'b': b'1', 'c': 4, 'd': 'qwe'})
-    assert m.td == {'a': 3, 'b': 1, 'c': 4, 'd': 'qwe'}
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(td={'a': [1], 'b': 2, 'c': 3, 'd': 'qwe'})
-    assert exc_info.value.errors() == [
-        {
-            'loc': ('td', 'a'),
-            'msg': 'value is not a valid integer',
-            'type': 'type_error.integer',
-        }
-    ]
-
-
-@pytest.mark.skipif(not TypedDict, reason='typing_extensions not installed')
-def test_typed_dict_non_total():
-    class FullMovie(TypedDict, total=True):
-        name: str
-        year: int
-
-    class Model(BaseModel):
-        movie: FullMovie
-
-    with pytest.raises(ValidationError) as exc_info:
-        Model(movie={'year': '2002'})
-    assert exc_info.value.errors() == [
-        {
-            'loc': ('movie', 'name'),
-            'msg': 'field required',
-            'type': 'value_error.missing',
-        }
-    ]
-
-    class PartialMovie(TypedDict, total=False):
-        name: str
-        year: int
-
-    class Model(BaseModel):
-        movie: PartialMovie
-
-    m = Model(movie={'year': '2002'})
-    assert m.movie == {'year': 2002}
